+++ frontend/src/views/MessstelleView.vue
@@ -54,10 +54,10 @@ import { computed, ComputedRef, ref, Ref } from "vue";
 import MessstelleService from "@/api/service/MessstelleService";
 import MessstelleInfoDTO from "@/types/MessstelleInfoDTO";
 import { useRoute } from "vue-router/composables";
-import MessstelleInfo from "@/components/messstelle/MessstelleInfo.vue";
-import MessquerschnittInfo from "@/components/messstelle/MessquerschnittInfo.vue";
+import MessstelleInfo from "@/components/messstelle/MessstelleHeaderInfo.vue";
+import MessquerschnittInfo from "@/components/messstelle/MessstelleInfo.vue";
 import { useVuetify } from "@/util/useVuetify";
-import MessungenTimeline from "@/components/messstelle/MessungenTimeline.vue";
+import MessungenTimeline from "@/components/messstelle/MessquerschnittInfo.vue";
 import DefaultObjectCreator from "@/util/DefaultObjectCreator";
 import { ApiError } from "@/api/error";
 import { useStore } from "@/api/util/useStore";
